requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
gspread==6.1.2
google-auth==2.31.0
